changed int to double

+++ src/main/java/com/helloworld/api/dto/PersonDTO.java
@@ -1,5 +1,5 @@
 package com.helloworld.api.dto;
 
-public record PersonDTO (String modelo, String fabricante, String dataFabricacao, int valor, int anoModelo) {
+public record PersonDTO (String modelo, String fabricante, String dataFabricacao, double valor, int anoModelo) {
     
 }
